requirements.txt
antlr4-python3-runtime
antlr4-tools
pytest